+++ User_File/Middleware/System/Init/Init.cpp
@@ -4,6 +4,7 @@
 #include "bsp_spi.h"
 #include "callback.h"
 #include "sys_timestamp.h"
+#include "bsp_can.h"
 
 // 全局初始化完成标志位
 volatile bool init_finished = false;
@@ -12,9 +13,12 @@ extern "C" void System_Init(void)
 {
     SEGGER_SYSVIEW_Conf();
     SYS_Timestamp.Init(&htim5); 
+
     HAL_NVIC_SetPriority(EXTI15_10_IRQn, 5, 0);
+
     // 陀螺仪的SPI
     SPI_Init(&hspi2, SPI2_Callback);
+
     // WS2812的SPI
     SPI_Init(&hspi6, nullptr);